project/requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-toolbelt==0.0.1
gunicorn==19.3.0
httplib2==0.9.1
psycopg2==2.6.1
python-instagram==1.3.2
simplejson==3.8.0
six==1.9.0
static3==0.6.1
wheel==0.24.0
